Prefer the last //# sourceMappingURL in a file instead of the first
It seems like some Webpack setups (including mine) end up outputting
several //# sourceMappingURL=... entries when both transpilation and
bundling is in the picture.

In general it seems like the safest bet to go for the last one, as
most tools tend to append the comment to the end of the file.
Thus the last one is most likely to pertain to the last operation
that happened, be it transpilation, bundling, or uglification :)